- exception refactored - ItemService.java, ItemRepository.java updated

+++ Sales/src/main/java/com/example/demo/model/repository/ItemRepository.java
@@ -2,18 +2,21 @@ package com.example.demo.model.repository;
 
 import com.example.demo.model.entity.Item;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import javax.transaction.Transactional;
+import javax.persistence.Query;
+import java.util.List;
 
-@Transactional
 @Repository
 public class ItemRepository {
 
     @PersistenceContext
     private EntityManager entityManager;
 
+    @Transactional(propagation = Propagation.MANDATORY)
     public Item addItem(Item item) {
 
         entityManager.persist(item);
@@ -21,22 +24,34 @@ public class ItemRepository {
         return item;
     }
 
+    @Transactional(propagation = Propagation.MANDATORY)
     public Item deleteItem(long itemId) {
 
         Item item = getItem(itemId);
 
-        entityManager.remove(item);
+        if(item != null)
+            entityManager.remove(item);
 
         return item;
     }
 
+    @Transactional(propagation = Propagation.SUPPORTS)
     public Item getItem(long itemId) {
 
         return entityManager.find(Item.class, itemId);
     }
 
+    @Transactional(propagation = Propagation.MANDATORY)
     public Item updateItem(Item item) {
 
         return entityManager.merge(item);
+    }
+
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public List<Item> getItemAll() {
+
+        Query query = entityManager.createQuery("Select i from Item i");
+
+        return query.getResultList();
     }
 }
